requirements.txt
dj_database_url==2.2.0
Django==5.1
django_allauth==64.1.0
django_heroku==0.3.1
fantasy_rankings_scraper==0.0.4
openai==1.43.0
pydantic==2.8.2
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.32
stripe==10.8.0
xhtml2pdf==0.2.16
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-debug-toolbar==4.2.0
PyJWT==2.6.0
gunicorn==21.2.0
pytz
django-extensions
sendgrid-django
django-storages
boto3
markdown2